ml/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
boto3==1.35.0
git+https://github.com/openai/CLIP.git
torch==2.4.1
torchvision==0.19.1
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.2
opencv-python-headless==4.10.0.84
anthropic>=0.40.0
